+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Added
 
 - Get one user with it's watched movies list.
-- todo: Get users list with it's corresponding watched movies.
+- Get users list with it's corresponding watched movies.
 - todo: Route that shows movies novelties list.
 
 # VERSION 0.3.0